Feature/crud open session api

+++ backend/src/main/java/com/jikateam/registration_course/dto/response/OpenSessionInfoResponse.java
@@ -10,7 +10,7 @@ public record OpenSessionInfoResponse(
         PhaseResponse registrationPhase,
         RegistrationStatus status,
         SessionInfoResponse session,
-        Integer numberOfRegister
+        Long numberOfRegister
 
 ) {
 }
